chore: split out async-rewriter tests into separate package
This will be a prerequisite for productionizing a new iteration
of the async rewriter, such as the one proposed in
https://github.com/mongodb-js/mongosh/pull/1931.
