+++ package.json
@@ -140,12 +140,13 @@
     "configs/eslint-config-mongosh",
     "configs/tsconfig-mongosh",
     "scripts/docker",
-    "packages/async-rewriter2",
+    "packages/async-rewriter-spec",
     "packages/build",
     "packages/errors",
     "packages/history",
     "packages/js-multiline-to-singleline",
     "packages/types",
+    "packages/async-rewriter2",
     "packages/i18n",
     "packages/logging",
     "packages/shell-bson",
