+++ packages/async-rewriter-spec/.eslintignore
@@ -1,0 +1,3 @@
+node_modules
+lib
+local-test.js
