+++ packages/async-rewriter-spec/.mocharc.json
@@ -1,0 +1,7 @@
+{
+  "require": ["../../scripts/import-expansions.js", "ts-node/register"],
+  "timeout": 60000,
+  "reporter": "../../configs/mocha-config-mongosh/reporter.ts",
+  "spec": ["./{src,lib}/**/*.spec.ts"],
+  "exclude": ["**/node_modules/**"]
+}
