+++ packages/async-rewriter-spec/.prettierrc.json
@@ -1,0 +1,1 @@
+"@mongodb-js/prettier-config-devtools"
