+++ packages/async-rewriter-spec/AUTHORS
@@ -1,0 +1,14 @@
+Anna Henningsen <anna@addaleax.net>
+github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
+Michael Rose <michael_rose@gmx.de>
+dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
+Maurizio Casimirri <maurizio.cas@gmail.com>
+kyriakos markakis <11089475+marquicodes@users.noreply.github.com>
+Alena Khineika <alena.khineika@gmail.com>
+Martin Rodriguez Reboredo <yakoyoku@gmail.com>
+Le Roux Bodenstein <lerouxb@gmail.com>
+Gagik Amaryan <me@gagik.co>
+mongodb-devtools-bot[bot] <189715634+mongodb-devtools-bot[bot]@users.noreply.github.com>
+Ubuntu <ubuntu@ip-10-0-5-160.ec2.internal>
+Ubuntu <ubuntu@ip-10-0-4-95.ec2.internal>
+Nikola Irinchev <irinchev@me.com>
