+++ packages/async-rewriter-spec/README.md
@@ -1,0 +1,5 @@
+# async-rewriter tests
+
+Re-usable test suite for mongosh's async rewriting engine.
+
+Details of the engine are described in the async-rewriter2 package README.
