+++ packages/async-rewriter-spec/package.json
@@ -1,0 +1,39 @@
+{
+  "name": "@mongosh/async-rewriter-spec",
+  "version": "1.0.0",
+  "description": "MongoDB Shell Async Rewriter Spec Package",
+  "license": "Apache-2.0",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "private": true,
+  "exports": {
+    ".": {
+      "default": "./src/index.ts"
+    }
+  },
+  "scripts": {
+    "eslint": "eslint",
+    "lint": "npm run eslint . && npm run prettier -- --check .",
+    "prettier": "prettier",
+    "reformat": "npm run prettier -- --write . && npm run eslint -- --fix",
+    "check": "npm run lint"
+  },
+  "config": {
+    "unsafe-perm": true
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/mongodb-js/mongosh.git"
+  },
+  "license": "Apache-2.0",
+  "engines": {
+    "node": ">=14.15.1"
+  },
+  "devDependencies": {
+    "@mongodb-js/eslint-config-mongosh": "^1.0.0",
+    "@mongodb-js/prettier-config-devtools": "^1.0.1",
+    "@mongodb-js/tsconfig-mongosh": "^1.0.0",
+    "eslint": "^7.25.0",
+    "prettier": "^2.8.8"
+  }
+}
