+++ packages/async-rewriter-spec/tsconfig-lint.json
@@ -1,0 +1,5 @@
+{
+  "extends": "./tsconfig.json",
+  "include": ["./src/**/*", "./benchmark/**/*", "./test/**/*"],
+  "exclude": ["node_modules", "dist", "lib"]
+}
