+++ packages/async-rewriter-spec/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@mongodb-js/tsconfig-mongosh/tsconfig.common.json",
+  "compilerOptions": {
+    "outDir": "./lib",
+    "allowJs": true
+  },
+  "include": ["src/**/*"],
+  "exclude": ["./src/**/*.spec.*"]
+}
