+++ packages/async-rewriter2/package.json
@@ -49,6 +49,7 @@
     "@mongodb-js/eslint-config-mongosh": "^1.0.0",
     "@mongodb-js/prettier-config-devtools": "^1.0.1",
     "@mongodb-js/tsconfig-mongosh": "^1.0.0",
+    "@mongosh/async-rewriter-spec": "^1.0.0",
     "@types/babel__core": "^7.20.5",
     "eslint": "^7.25.0",
     "prettier": "^2.8.8"
