+++ packages/async-rewriter2/src/index.ts
@@ -1,2 +1,3 @@
 import AsyncWriter from './async-writer-babel';
+export { AsyncWriter };
 export default AsyncWriter;
